refactor: modularize settings dialog sections
