+++ src/components/settings/index.ts
@@ -1,0 +1,4 @@
+export { DevicePermissionsSection } from './device-permissions-section';
+export { GeneralAccessSection } from './general-access-section';
+export { PlaceholderSection } from './placeholder-section';
+export { ThemeSection } from './theme-section';
